requirements.txt
# 基础依赖
numpy>=1.20.0
pillow>=8.0.0
pyyaml>=6.0
requests>=2.25.0
tqdm>=4.60.0
python-dotenv>=0.19.0
opencv-python>=4.5.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-image>=0.18.0
scikit-learn>=0.24.0

# 图像处理
torch>=1.9.0
torchvision>=0.10.0
segment-anything>=1.0  # https://github.com/facebookresearch/segment-anything
rembg>=2.0.0  # 背景移除
photoshop-python-api>=0.15.0  # PSD文件处理
psd-tools>=1.9.0  # 替代pypsd的更现代PSD处理库
pymatting>=1.1.5  # Alpha matting

# AI模型接口
diffusers>=0.11.0
transformers>=4.18.0
accelerate>=0.17.0
controlnet-aux>=0.0.6
# xformers>=0.0.17  # 可选，用于优化注意力机制

# UI自动化
pyautogui>=0.9.53
pynput>=1.7.6
mss>=6.1.0  # 屏幕截图
pytesseract>=0.3.9  # OCR
pywinctl>=0.0.46  # 窗口控制 (Windows)
pygetwindow>=0.0.9

# 日志和调试
loguru>=0.6.0
rich>=12.0.0

# 系统和文件处理
psutil>=5.9.0
python-magic>=0.4.24; platform_system != "Windows"  # 文件类型检测
python-magic-bin>=0.4.14; platform_system == "Windows"
zipfile38>=0.0.3

# 测试和开发
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.3.0
isort>=5.10.0
mypy>=0.950
types-requests>=2.27.25
types-PyYAML>=6.0.0

# 文档
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.8

# 自动化和UI
pyautogui==0.9.54
pynput==1.7.6
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1

# API调用
httpx>=0.24.0  # 替代aiohttp的现代HTTP客户端库

# GUI
gradio>=3.40.1
fastapi>=0.98.0
uvicorn>=0.22.0
websockets>=11.0.3
jinja2>=3.1.2
ffmpy>=0.3.1
python-multipart>=0.0.6
httpx>=0.24.1
pydub>=0.25.1
orjson>=3.9.1 